+++ archsim/modules/arm-system-devices/inc/SP810.h
@@ -29,8 +29,8 @@ namespace archsim
 
 				bool Initialise() override;
 
-				bool Read(uint32_t offset, uint8_t size, uint32_t& data) override;
-				bool Write(uint32_t offset, uint8_t size, uint32_t data) override;
+				bool Read(uint32_t offset, uint8_t size, uint64_t& data) override;
+				bool Write(uint32_t offset, uint8_t size, uint64_t data) override;
 
 				uint64_t hr_begin;
